clean up test matrix
