+++ .github/workflows/test.yml
@@ -6,9 +6,16 @@ on:
   pull_request:
 
 jobs:
-  test-rs:
+  test:
     runs-on: ubuntu-22.04
-    name: Test Replica Sets
+    name: Run ${{ matrix.options[0] }} Tests
+    matrix:
+        # Github Actions doesn't support pairing matrix values together, let's improvise
+        # https://github.com/github/feedback/discussions/7835#discussioncomment-1769026
+        options:
+        - ["Other", "--ignore=tests/test_replica_sets.py --ignore=tests/test_replica_sets2.py"]
+        - ["Replica Set", "tests/test_replica_set.py"]
+        - ["Replica Sets", "tests/test_replica_sets.py"]
     steps:
       - uses: actions/checkout@v3
       - name: Setup Python
@@ -28,28 +35,4 @@ jobs:
           pip install pytest
       - name: Run Tests
         run: |
-          pytest -raXs -v --durations 10 --color=yes tests/test_replica_sets.py
-
-  test-other:
-    runs-on: ubuntu-22.04
-    name: Test Other Files
-    steps:
-      - uses: actions/checkout@v3
-      - name: Setup Python
-        uses: actions/setup-python@v2
-        with:
-          cache: 'pip'
-          cache-dependency-path: 'setup.py'
-      - name: Install MongoDB
-        run: |
-          wget -qO - https://www.mongodb.org/static/pgp/server-6.0.asc | sudo apt-key add -
-          echo "deb [ arch=amd64,arm64 ] https://repo.mongodb.org/apt/ubuntu jammy/mongodb-org/6.0 multiverse" | sudo tee /etc/apt/sources.list.d/mongodb-org-6.0.list
-          sudo apt-get update
-          sudo apt-get install -y mongodb-org
-      - name: Install deps
-        run: |
-          pip install -e .
-          pip install pytest
-      - name: Run Tests
-        run: |
-          pytest -raXs -v --durations 10 --color=yes --ignore=tests/test_replica_sets.py
+          pytest -raXs -v --durations 10 --color=yes ${{ matrix.options[1] }}
